+++ frb_codegen/src/library/codegen/parser/mir/parser/mod.rs
@@ -1,5 +1,6 @@
 pub(crate) mod attribute;
 pub(crate) mod auto_accessor;
 pub(crate) mod function;
+pub(crate) mod misc;
 pub(crate) mod trait_impl;
 pub(crate) mod ty;
