+++ frb_codegen/src/library/codegen/parser/mir/sanity_checker/unused_checker.rs
@@ -4,7 +4,7 @@ use crate::codegen::ir::mir::pack::MirPack;
 use crate::codegen::ir::mir::ty::delegate::MirTypeDelegate;
 use crate::codegen::ir::mir::ty::MirType;
 use crate::codegen::parser::mir::internal_config::RustInputNamespacePack;
-use crate::codegen::parser::mir::misc::extract_src_types_in_paths;
+use crate::codegen::parser::mir::parser::misc::extract_src_types_in_paths;
 use crate::codegen::parser::mir::parser::ty::path_data::extract_path_data;
 use crate::utils::namespace::NamespacedName;
 use itertools::Itertools;
